Feat(ServiceWorkerFlow): Caminho para quando o service worker é registrado pela primeira vez

+++ src/components/sync_info/index.tsx
@@ -7,7 +7,10 @@ import SyncService from '../../services/sync/SyncService'
 import SyncCloudProps from './props'
 import './styles.css'
 
-const SyncInfo: React.FC<SyncCloudProps> = ({ className }) => {
+const SyncInfo: React.FC<SyncCloudProps> = ({
+	className,
+	state 
+}) => {
 	const [isLoading, setIsLoading] = useState(true)
 	const [syncState, setSyncState] = useState(SyncStateEnum.SYNCHRONIZING)
 
@@ -18,14 +21,18 @@ const SyncInfo: React.FC<SyncCloudProps> = ({ className }) => {
 			setIsLoading(false)
 		}
 
-		SyncService.addUpdateEventListenner(loadData)
+		if (state === undefined) SyncService.addUpdateEventListenner(loadData)
 
 		if (isLoading) {
-			loadData()
+			if (state === undefined) {
+				loadData()
+			} else {
+				setSyncState(state)
+			}
 		}
 
 		return () => {
-			SyncService.removeUpdateEventListenner(loadData)
+			if (state === undefined) SyncService.removeUpdateEventListenner(loadData)
 		}
 	}, [isLoading])
